new task added

+++ src/Controller/ProjectController.php
@@ -3,28 +3,27 @@
 namespace App\Controller;
 
 use App\Entity\Project;
-use App\Form\ProjectType;
+use App\Form\Project1Type;
 use App\Repository\ProjectRepository;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
 
-#[Route('/project')]
+
 
 /**
  * Class ProjectController
  * @package App\Controller
- * @Route("/project", name="project")
+ * @Route("/project")
  */
 class ProjectController extends AbstractController
 {
 
-
     /**
      * @param ProjectRepository $projectRepository
      * @return Response
-     * @Route("/", name="project_index")
+     * @Route("/", name="project_index", methods={"GET"})
      */
     public function index(ProjectRepository $projectRepository): Response
     {
@@ -33,11 +32,17 @@ class ProjectController extends AbstractController
         ]);
     }
 
-    #[Route('/new', name: 'project_new', methods: ['GET', 'POST'])]
+
+
+    /**
+     * @param Request $request
+     * @return Response
+     * @Route("/new", name="project_new", methods={"GET","POST"})
+     */
     public function new(Request $request): Response
     {
         $project = new Project();
-        $form = $this->createForm(ProjectType::class, $project);
+        $form = $this->createForm(Project1Type::class, $project);
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
@@ -54,12 +59,12 @@ class ProjectController extends AbstractController
         ]);
     }
 
-    #[Route('/{id}', name: 'project_show', methods: ['GET'])]
+
 
     /**
      * @param Project $project
      * @return Response
-     * @Route("/{id}", name="project_show")
+     * @Route("/{id}", name="project_show", methods={"GET"})
      */
     public function show(Project $project): Response
     {
@@ -68,10 +73,17 @@ class ProjectController extends AbstractController
         ]);
     }
 
-    #[Route('/{id}/edit', name: 'project_edit', methods: ['GET', 'POST'])]
+
+
+    /**
+     * @param Request $request
+     * @param Project $project
+     * @return Response
+     * @Route("/{id}/edit", name="project_edit", methods={"GET","POST"})
+     */
     public function edit(Request $request, Project $project): Response
     {
-        $form = $this->createForm(ProjectType::class, $project);
+        $form = $this->createForm(Project1Type::class, $project);
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
@@ -86,15 +98,30 @@ class ProjectController extends AbstractController
         ]);
     }
 
-    #[Route('/del/{id}', name: 'project_delete', methods: ['DELETE'])]
+
+    /**
+     * @param Request $request
+     * @param Project $project
+     * @return Response
+     * @Route("/{id}", name="project_delete", methods={"DELETE"})
+     */
     public function delete(Request $request, Project $project): Response
     {
-
+        if ($this->isCsrfTokenValid('delete'.$project->getId(), $request->request->get('_token'))) {
             $entityManager = $this->getDoctrine()->getManager();
             $entityManager->remove($project);
             $entityManager->flush();
-
+        }
 
         return $this->redirectToRoute('project_index');
+    }
+
+    /**
+     * @return Response
+     * @Route("/hej")
+     */
+    public function showTasks():Response
+    {
+        return new Response("hello wordl");
     }
 }
